+++ apps/docs/src/content/03-components/status/alert-text/examples/default.tsx
@@ -1,0 +1,3 @@
+import { AlertText } from "@mittwald/flow-react-components";
+
+<AlertText>Info</AlertText>;