+++ apps/docs/src/content/03-components/status/alert-text/examples/status.tsx
@@ -1,0 +1,9 @@
+import { AlertText } from "@mittwald/flow-react-components";
+
+<Row>
+  <AlertText status="info">Info</AlertText>
+  <AlertText status="warning">Warning</AlertText>
+  <AlertText status="danger">Danger</AlertText>
+  <AlertText status="success">Success</AlertText>
+  <AlertText status="unavailable">Unavailable</AlertText>
+</Row>;