+++ packages/remote-react-components/src/tests/visual/CartesianChart.browser.test.tsx
@@ -42,7 +42,7 @@ test.each(testEnvironments)(
       >
         <CartesianGrid />
         <Area dataKey="first" />
-        <Area dataKey="second" color="palatinate-blue" />
+        <Area dataKey="second" color="#555" />
         <Line dataKey="third" color="magenta" />
         <XAxis dataKey="name" />
         <YAxis />