+++ packages/remote-react-components/src/tests/visual/DonutChart.browser.test.tsx
@@ -46,9 +46,10 @@ test.each(testEnvironments)(
             key={position}
             aria-label="label"
             segments={[
-              { title: "Segment 1", value: 50 },
-              { title: "Segment 2", value: 25 },
-              { title: "Segment 3", value: 12 },
+              { title: "Segment 1", value: 30 },
+              { title: "Segment 2", value: 15 },
+              { title: "Segment 3", value: 25, color: "lime" },
+              { title: "Segment 4", value: 12, color: "#555" },
             ]}
             legendPosition={position}
           />