+++ packages/remote-react-components/src/tests/visual/Message.browser.test.tsx
@@ -20,6 +20,7 @@ test.each(testEnvironments)(
       Text,
       FileCard,
       FileCardList,
+      ActionGroup,
     },
   }) => {
     await render(
@@ -53,6 +54,12 @@ test.each(testEnvironments)(
               </FileCardList>
             </Flex>
           </Content>
+          <ActionGroup>
+            <Button variant="soft" color="secondary">
+              Secondary
+            </Button>
+            <Button>Primary</Button>
+          </ActionGroup>
         </Message>
         <Message type="sender">
           <Content>
@@ -62,6 +69,13 @@ test.each(testEnvironments)(
               accusantium rerum ullam modi blanditiis vitae, laborum ea tempore,
               dolore voluptas.
             </Text>
+          </Content>
+
+          <Button>Button</Button>
+        </Message>
+        <Message color="#ffeedd">
+          <Content>
+            <Text>Custom color</Text>
           </Content>
         </Message>
       </Flex>,